new checkpoints file. tweaked colors. Tried fixing Delete on send enabled on zero records.

+++ src/main/java/org/multibit/viewsystem/swing/view/panels/SendBitcoinPanel.java
@@ -90,7 +90,11 @@ public class SendBitcoinPanel extends AbstractTradePanel implements Viewable {
 
     @Override
     protected Action getDeleteAddressAction() {
-        return new DeleteSendingAddressAction(controller, mainFrame, this);
+        if (deleteAddressAction == null) {
+            return new DeleteSendingAddressAction(controller, mainFrame, this);
+        } else {
+            return deleteAddressAction;
+        }
     }
     
     @Override
